Don't advertise homekit service until config init
Previously, the port defined in homekit.xml was being passed on init, and we started the service advertisement there. Then, when the configuration was processed and passed to us, we'd create a new instance of the service. We now shutdown any existing services before starting, and prevent the initial useless advertisement, which was causing iOS to delay its reconnect.